refactor(fab, button): update imports and restructure components
Removed the index file from the Fab component and adjusted all related imports to point directly to their respective files. Streamlined the directory structure for Fab and Button components for improved maintainability and consistency.
