--- src/components/action/fab/index.ts
@@ -1,3 +1,0 @@
-export * from '../Fab';
-export * from '../../../interfaces/fab.interface';
-export * from '../../../styles/fab.style';
